MAYA114898 Merge to USD options dialog
- Add a exportOptions flag to the merge-to-USD and duplicate commands.
- Refactor the code in the export translator that parses the options string to be in a reusable function.
- Add it as GetDictionaryFromEncodedOptions in UsdMayaJobExportArgs.
- Add the merge-to-USD options dialog UI.

Still todo: extracting the options, setting up the default options, some UI layout work.
