+++ lib/mayaUsd/fileio/jobs/jobArgs.h
@@ -249,6 +249,20 @@ struct UsdMayaJobExportArgs
         const UsdMayaUtil::MDagPathSet& dagPaths,
         const std::vector<double>&      timeSamples = std::vector<double>());
 
+    /// Creates a UsdMayaJobExportArgs from the given text-encoded options,
+    /// overlaid on top of the default dictionary given by GetDefaultDictionary().
+    /// The values of text-encoded options are stronger (will override)
+    /// the values from the default dictionary.
+    /// Issues runtime errors if some options contains values of the wrong type;
+    /// types should match those declared in GetDefaultDictionary().
+    ///
+    /// The text encoding is in the form: name1=value1;name2=value2;...
+    MAYAUSD_CORE_PUBLIC
+    static MStatus GetDictionaryFromEncodedOptions(
+        const MString&       textEncodedOptions,
+        VtDictionary*        userArgs,
+        std::vector<double>* timeSamples);
+
     /// Gets the default arguments dictionary for UsdMayaJobExportArgs.
     MAYAUSD_CORE_PUBLIC
     static const VtDictionary& GetDefaultDictionary();
